requirements.txt
# matrx          == 2.2.0

# all requirements below are 
# additional libraries provided to do this assignment. 


# libraries used by matrx and their versions, excl webserver stuff
bidict        >= 0.21.2
chardet       >= 4.0.0
colour        >= 0.1.5
docutils       >= 0.16
gevent         >= 21.1.2
greenlet       >= 3.0.0rc3
jsonpickle     >= 1.5.1
numpy          >= 1.20.0
Werkzeug       >= 2.0.0
flask          >= 2.1.3
pyyaml
ollama         >= 0.4.0   # Ollama Python SDK for tool calling (local mode)
requests       >= 2.28.0  # HTTP client (used for direct Ollama REST API calls)
transformers   >= 4.51.0  # HuggingFace transformers for in-process LLM (HPC mode)
torch          >= 2.1.0   # PyTorch backend for transformers
accelerate     >= 0.26.0  # Required for device_map="auto" in from_pretrained
toon-format  # Token-Oriented Object Notation: 30-60% token reduction for LLM inputs
langgraph     >= 0.2.0   # LangGraph for graph-based agent workflows
langchain-core >= 0.3.0  # Required by langgraph (Runnable interface)
